+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/yaml/pojo/YamlRegistryCenterConfiguration.java
@@ -15,20 +15,24 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.governance.core.yaml.config.pojo.scaling;
+package org.apache.shardingsphere.governance.core.yaml.pojo;
 
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.shardingsphere.infra.yaml.config.pojo.YamlConfiguration;
 
+import java.util.Properties;
+
 /**
- * Scaling configuration for YAML.
+ * Registry center configuration for YAML.
  */
 @Getter
 @Setter
-public final class YamlScalingConfiguration implements YamlConfiguration {
+public final class YamlRegistryCenterConfiguration implements YamlConfiguration {
     
-    private int blockQueueSize;
+    private String type;
     
-    private int workerThread;
+    private String serverLists;
+    
+    private Properties props = new Properties();
 }
