+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-governance-spring/shardingsphere-jdbc-governance-spring-boot-starter/src/main/java/org/apache/shardingsphere/spring/boot/governance/common/GovernanceSpringBootRootConfiguration.java
@@ -19,7 +19,7 @@ package org.apache.shardingsphere.spring.boot.governance.common;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.apache.shardingsphere.governance.core.yaml.config.pojo.YamlGovernanceConfiguration;
+import org.apache.shardingsphere.governance.core.yaml.pojo.YamlGovernanceConfiguration;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 import java.util.Properties;
