+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-governance-spring/shardingsphere-jdbc-governance-spring-boot-starter/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -77,7 +77,7 @@
     },
     {
       "name": "spring.shardingsphere.governance",
-      "type": "java.util.Map<java.lang.String,org.apache.shardingsphere.governance.core.yaml.config.pojo.YamlRegistryCenterConfiguration>",
+      "type": "java.util.Map<java.lang.String,org.apache.shardingsphere.governance.core.yaml.pojo.YamlRegistryCenterConfiguration>",
       "sourceType": "org.apache.shardingsphere.spring.boot.governance.common.GovernanceSpringBootRootConfiguration",
       "description": "Customize ShardingSphere governance instance."
     },
